refrac(datasets): use simple factory to manage datasets

+++ datasets/mmoe_synthetic.py
@@ -6,7 +6,7 @@ class SynDataset(torch.utils.data.Dataset):
     
     def __init__(self, dataset_path):
         data = pd.read_csv(dataset_path).to_numpy()[:, 0:]
-        self.categorical_data = data[:, 0:0].astype(np.int)
+        self.categorical_data = data[:, 0:0].astype(np.int32)
         self.numerical_data = data[:, : -2].astype(np.float32)
         self.labels = data[:, -2:].astype(np.float32)
         self.numerical_num = self.numerical_data.shape[1]
